number_1/requirements.txt
apache-airflow[postgres]==2.7.0
psycopg2-binary==2.9.9
pandas==2.0.3
sqlalchemy==1.4.49
